+++ src/update/main.cpp
@@ -34,6 +34,7 @@
 #include "UpdateTo_0.8.5.hpp"
 #include "UpdateTo_0.8.6.hpp"
 #include "UpdateTo_0.9.0.hpp"
+#include "UpdateTo_0.9.7.hpp"
 
 void showVersion()
 {
@@ -249,6 +250,15 @@ int main(int argc, char** argv)
   {
     std::cout << "Update for database of version 0.8.9 (and earlier) to version 0.9.0..." << std::endl;
     if (!wic::UpdateTo090::perform(config.connectionInfo()))
+    {
+      std::cerr << "Error: Database update failed!" << std::endl;
+      return wic::rcUpdateFailure;
+    }
+  }
+  if (currentVersion < wic::SemVer(0, 9, 7))
+  {
+    std::cout << "Update for database of version 0.9.6 (and earlier) to version 0.9.7..." << std::endl;
+    if (!wic::UpdateTo097::perform(config.connectionInfo()))
     {
       std::cerr << "Error: Database update failed!" << std::endl;
       return wic::rcUpdateFailure;