+++ tests/general/api/Types.cpp
@@ -55,6 +55,14 @@ TEST_CASE("API types")
     REQUIRE( wic::toApiType("\t  WeAtHeRbIt  ") == wic::ApiType::Weatherbit );
   }
 
+  SECTION("to API type: Weatherstack")
+  {
+    REQUIRE( wic::toApiType("weatherstack") == wic::ApiType::Weatherstack );
+    REQUIRE( wic::toApiType("Weatherstack") == wic::ApiType::Weatherstack );
+    REQUIRE( wic::toApiType("WEATHERSTACK") == wic::ApiType::Weatherstack );
+    REQUIRE( wic::toApiType("\t  WeAtHeRsTaCk  ") == wic::ApiType::Weatherstack );
+  }
+
   SECTION("to API type: none")
   {
     REQUIRE( wic::toApiType("none") == wic::ApiType::none );
@@ -77,6 +85,7 @@ TEST_CASE("API types")
     REQUIRE( wic::toString(wic::ApiType::OpenWeatherMap) == "OpenWeatherMap" );
     REQUIRE( wic::toString(wic::ApiType::DarkSky) == "DarkSky" );
     REQUIRE( wic::toString(wic::ApiType::Weatherbit) == "Weatherbit" );
+    REQUIRE( wic::toString(wic::ApiType::Weatherstack) == "Weatherstack" );
     REQUIRE( wic::toString(wic::ApiType::none) == "none" );
   }
 
@@ -86,6 +95,7 @@ TEST_CASE("API types")
     REQUIRE( wic::toApiType(wic::toString(wic::ApiType::OpenWeatherMap)) == wic::ApiType::OpenWeatherMap );
     REQUIRE( wic::toApiType(wic::toString(wic::ApiType::DarkSky)) == wic::ApiType::DarkSky );
     REQUIRE( wic::toApiType(wic::toString(wic::ApiType::Weatherbit)) == wic::ApiType::Weatherbit );
+    REQUIRE( wic::toApiType(wic::toString(wic::ApiType::Weatherstack)) == wic::ApiType::Weatherstack );
     REQUIRE( wic::toApiType(wic::toString(wic::ApiType::none)) == wic::ApiType::none );
   }
 }